added basic tests fixed infinite recursion bug added toString methods to logs and stats objects

+++ src/org/mondemand/Client.java
@@ -363,7 +363,7 @@ public class Client {
      * @param value the value to set this counter to
      */
     public void setKey(String key, int value) {
-    	setKey(key, value);
+    	setKey(key, (long) value);
     }
     
     /**
